Remove fixture-dependent and VP2-dependent tests from unit/integration suites
- Delete Lumine PMX dependent tests (8 tests across 4 files)
- Delete PMD fixture dependent tests (test_pmd_importer.py + 5 tests in 4 files)
- Delete stale VMD/PMD exporter test files (de437e4)
- Move VP2/GUI shader override tests to tests/gui/guitest_shader_override.py
- Fix test_import_pmx_basic to use mmt_test_model instead of Lumine
- Simplify _import_model_and_apply_vmd helper (PMX only)
- Remove unused PmdData import and dead pmd_file references

Unit: 794 pass, 3 skip, 0 fail
Integration: 100 pass, 4 skip, 0 fail
