+++ tests/integration/test_bone_converter.py
@@ -30,73 +30,6 @@ class TestBoneConverter(MayaTestBase):
         cmds.file(new=True, force=True)
         self.fixture_provider.cleanup_temp_files()
 
-    def test_convert_pmd_bones(self):
-        """PMDボーンがMayaに正しく変換され、スキニングが適用されることをテストする。"""
-        # PMDファイルが存在するか確認
-
-        pmd_data, pmd_file_path = self.fixture_provider.load_pmd_data("miku_v2")
-
-        root_group = cmds.group(name="pmd_model_root", empty=True)
-
-        # テスト用のメッシュを作成
-        pmd_mesh_converter = MeshConverter(pmd_file_path)
-        pmd_group_name, pmd_mesh_name = pmd_mesh_converter.convert_pmd_mesh(pmd_data, root_group)
-
-        # ボーンを変換
-        converter = BoneConverter()
-        root_joint, skin_cluster = converter.convert_pmd_bones(pmd_data, pmd_mesh_name, root_group)
-
-        # 結果を検証
-        self.assertIsNotNone(root_joint, "ルートジョイントが作成されていません。")
-        self.assertIsNotNone(skin_cluster, "スキンクラスターが作成されていません。")
-
-        # ジョイントの数を確認（準標準ボーンが追加される可能性がある）
-        all_joints = cmds.ls(type="joint")
-        # PMDボーン数以上のジョイントが作成されることを確認
-        self.assertGreaterEqual(
-            len(all_joints),
-            len(pmd_data.bones),
-            "ジョイント数が元のPMDボーン数以上でなければなりません。",
-        )
-
-        # 階層構造と位置を確認
-        for bone in pmd_data.bones:
-            bone_name = maya_utils.sanitize_text(bone.get_name())  # 英語名があればそれを使用
-            self.assertTrue(
-                cmds.objExists(bone_name),
-                f"ジョイント '{bone_name}' が作成されていません。",
-            )
-
-            # 親子関係の確認
-            if bone.parent_bone_index != -1:
-                parent_name = pmd_data.bones[bone.parent_bone_index].get_name()
-                parent_name = maya_utils.sanitize_text(parent_name)
-                parent_joint = cmds.listRelatives(bone_name, parent=True, type="joint")
-                self.assertIsNotNone(parent_joint, f"ジョイント '{bone_name}' に親がいません。")
-
-            # ジョイントが存在する場合のみ位置を確認
-            if cmds.objExists(bone_name):
-                # 位置の確認
-                joint_pos = cmds.xform(bone_name, query=True, translation=True, worldSpace=True)
-                self.assertAlmostEqual(
-                    joint_pos[0],
-                    bone.position[0],
-                    delta=1e-5,
-                    msg=f"ジョイント '{bone_name}' のX位置が正しくありません。",
-                )
-                self.assertAlmostEqual(
-                    joint_pos[1],
-                    bone.position[1],
-                    delta=1e-5,
-                    msg=f"ジョイント '{bone_name}' のY位置が正しくありません。",
-                )
-                self.assertAlmostEqual(
-                    joint_pos[2],
-                    -bone.position[2],
-                    delta=1e-5,
-                    msg=f"ジョイント '{bone_name}' のZ位置が正しくありません。",
-                )  # Mayaは左手系
-
     def test_convert_pmx_bones(self):
         """PMXボーンがMayaに正しく変換され、スキニングが適用されることをテストする。"""
         # PMXファイルをロード
